+++ compiler/AST/Makefile.share
@@ -32,6 +32,7 @@ AST_SRCS =                                          \
            flags.cpp                                \
            FnSymbol.cpp                             \
            ForallStmt.cpp                           \
+	   IfExpr.cpp			            \
            iterator.cpp                             \
            ModuleSymbol.cpp                         \
            PartialCopyData.cpp                      \
